merge callRemote-3
Author: exarkun
Reviewer: lots of people

Athena provides an alternate implementation of the LivePage concepts.
Rather than pushing javascript code between client and server, JSON-serialized
data is passed back and forth.  Each side is given a `callRemote` API (to be
expanded) which can invoke methods on the other (with arguments) and which
returns a Deferred which fires when the results are available.

Future plans include the addition of explicit Referenceables (a la PB) and
the further separation of method dispatch logic from the (Live)Page subclass.