Athena is awesome.  This commit is cool, and by cool i mean totally sweet.--More--
Re #411

This branch adds several features to Athena.

 - Divmod.Class

This object adds traditional class-based object-oriented inheritance to
JavaScript.  The features here are mainly related to the behavior of
constructors, which on Divmod.Class is a method called "__init__" rather than
the function which also represents the class.  "__init__" isn't called when
creating an instance for use in a subclass relationship, so your constructors
can actually require arguments.  If you don't override "__init__", your
superclass's constructor implementation can be used.  Finally, a convenience
method for interacting with inherited methods is provided, the "upcall" method.

 - Nevow.Athena.Widget

In Athena, pages can be broken up on the server side into LiveFragments.  On
the client, they can now be broken down into Widgets, which have a 1:1
correspondence with LiveFragments on the server, and are automatically created
by a template pass.  Widget is a Divmod.Class, so it can be conveniently
subclassed by user code.  See examples/athenademo/widgets.{py,js} for examples.

 - LiveFragment.callRemote

Previously only page had a callRemote method.  Now, LiveFragments can send
method calls directly to their corresponding Widget on the client, rather than
to global functions.  This is useful in particular because now it is possible
to place multiple LiveFragments of the same type on a page and have the
messages from the server's view matched up to the client's.

 - Server-to-Client call exception handling (i.e. the client reports exceptions
   to the server.)

When using LiveFragment.callRemote and Page.callRemote, exceptions are now
propogated into Python as Failure objects without tracebacks.  (Note:
Javascript *does* provide enough information to fill out the traceback
information on that failure, and it would be helpful if some intrepid soul
would do that.)