Simplify connection-tracking strategy by removing 10 dictionaries from the factory.
Author: mg, exarkun, glyph
Reviewer: Are you kidding there are enough authors already
Fixes #410
Re #344

This merge fixes 2 problems.  The first is that Athena pages would time out
after the first idle timeout; 'assert req is gotReq' tracebacks would be
displayed and the client would be unable to communicate further.

There is also an issue with Athena leaking various kinds of objects in
state-tracking dictionaries, and those dictionaries have now all been removed
in favor of attributes on the LivePage that are set during a '_becomeLive'
method called in renderHTTP prior to rendering the page.

There is still a traceback on the server when the client blocks and the
javascript breaks; for example, if you use alert().  Aside from UI and proper
log messages (rather than explosions in the log) this behavior is largely
correct.