Divmod.Defer.Deferred - implementation of Failure and Deferred using Divmod.Class.
This does error reporting a bit more zealously than the MochiKit.Async
implementation of Deferred.  It is quite a lot shorter, and integrated with the
Divmod module system, which makes it quite a lot easier to instrument when
things start to go pear-shaped.  That instrumentation is useful since Athena's
transport logic is pretty much permanently the shape of a pear right now.

Author: exarkun
Reviewer: glyph

Fixes #584