Merge faster-athena-963
Author: exarkun, glyph
Reviewer: moea, amberite, other people
Fixes #963

This removes _walkDOM and replaces usage of it with Divmod.Runtime.Platform.traverse,
which is significantly more performant.  It also adds tests for Widget.get() and for
the ordering of Platform.traverse().  Widgets are also looked up by their XHTML id
attribute, rather than an id attribute in the Athena XML namespace, which is also
significantly faster.