Merge js-modules-651.
Fixes #651
Author: mithrandi
Reviewer: exarkun

This branch implements {{{athena.AutoJSPackage}}} for use in nevow plugins; it
automatically derives JS module mappings from the filesystem hierarchy, in a
manner similar to Python. Unlike python, __init__.js files are optional, and
will be treated like an empty file if not present. Mantissa and Quotient are
also migrated to use the new mechanism.