Implement nodeInserted callback for Athena.
Fixes #1680
Author: mithrandi
Reviewer: exarkun

Athena will now invoke the nodeInserted method (if it exists) on a widget that
it instantiates statically; for dynamically instantiated widgets, the code
instantiating it is responsible for calling this. This eases implementing
widgets that need to ensure that certain "setup" tasks (eg. DOM manipulation)
are only performed once the widget node has been inserted into the browser
document element hierarchy.