Scrolling view for Hyperbola, plus various ScrollTable updates.
Author: glyph, moe

Reviewer: exarkun

Fixes #2156
Fixes #1828

This change adds a scrolltable to the Hyperbola blurb viewing page.  That allows larger numbers of posts to be displayed on, for example, a blog.

However, the scrolltable as developed in #673 wasn't quite up to the task, so this change also includes several fixes to the {{{ScrollTable}}} itself, including:

    * a timestamp column, which can be used with {{{epsilon.extime.Time}}} instances, and allows data to be sorted by time.

    * a client-side 'widget' column, allowing a scrolltable to contain a column that renders a complex server-side widget inline and have it connected properly

    * better plugabbility for the IColumn interface, including custom unserialization for the sorting logic in Python

    * bugfixes, and better test coverage, for descending sort ordering

    * bugfixes and documentation for misuse of the API (for example, not providing a sort column).

The outcome of this group of changes is that the new {{{ScrollTable}}} is now an API that could reasonably be used in a variety of applications that require a scrolling list of data.  It isn't ''finished'', some applications will still require additional features which we have not added yet (for example, sort column header rendering) but others will simply be able to use it directly now.