Merge liveform-input-error-2169-2
Author: exarkun
Reviewer: glyph
Fixes #2169

Add an InputError exception to Mantissa LiveForm which can be used by
server-side components to indicate an input verification failure to
the client.  Also add logic to the client to handle this kind of error
specially.

Also fix a bug in MockBrowser to allow TextNodes to be added as children
to other nodes and fix a bug in removeChild so that parentNode is not
set after a child node has been removed from its parent.