Implement unit tests Athena XHR and correct some invalid assumptions about browsers.
Author: glyph

Fixes #2206

Fixes #1613

Reviewer: moe

This change implements a bunch of new JavaScript unit tests for both the Runtime's getPage method, and the ReliableMessageDelivery object.  In so doing, various parts of the process are refactored, and there is much less reliance on global state now.  In particular, the _actionHandlers global object literal bag of functions was translated into a set of regular methods on the ReliableMessageDelivery object.  This change does not introduce ''complete'' coverage, nor does it refactor all the unnecessary reliance upon global state, but it does provide enough coverage that we may begin to change the message queue behavior with some level of confidence.

Since there wouldn't have been much of a point to writing tests to verify behavior which was known to be incorrect at the time, we also change the behavior of the 'onunload' handler here to be in 'onbeforeunload', and use a synchronous request, as asynchronous requests will not be honored in this context (except due to a bug in some Firefox extensions that wrap XHR, such as Firebug and some versions of YSlow), and to clear the queue and use a special sequence number to block the user's browser as little as possible.  The accompanying Python change (to recognize the special meaning of 'unload' as a sequence number) is also implemented here, along with tests.