Apply patch removing broken support code for stackless python
Author: exarkun
Reviewer: glyph
Fixes: #2610

Nevow tried to register adapters for stackless-only types as
part of the package import.  Stackless no longer has these
types anymore, so with this code Nevow was completely unusable
on stackless.  There was no test coverage for this code, so
we'll just get rid of it and perhaps later on an interested
party can supply a patch to add special stackless features (if
it even makes sense to do so anymore).