Fixing athena unit tests.
To enable that, re-adding a json module.

In contrast to the original nevow.json, this is a wrapper around the built-in
json module with extra behaviour for nevow classes; also, this uses loads and
dumps rather than parse and serialize as before for consistency with builtin
json.

Dropped support for unsupported in json, dropped support for
embedding executable javascript in json.  I suspect we'll need
this if people actually use athena.  Does anyone?

Json will no longer interpret \x escapes (it's not RFC 8259, and I hope
it's not used anywhere).

Consequently, restored and fixed test_json.